used oauth2 for passport stragey of google

+++ controllers/api/v1/posts_api.js
@@ -28,20 +28,19 @@ module.exports.destroy = async function(req, res){
     try{
         let post = await Post.findById(req.params.id);
 
-        // if (post.user == req.user.id){
+        if (post.user == req.user.id){
             post.remove();
 
             await Comment.deleteMany({post: req.params.id});
 
-
-    
             return res.json(200, {
                 message: "Post and associated comments deleted successfully!"
             });
-        // }else{
-        //     req.flash('error', 'You cannot delete this post!');
-        //     return res.redirect('back');
-        // }
+        }else{
+            return res.json(401,{
+               message:"you can not delete this post"
+            });
+        }
 
     }catch(err){
         console.log('********', err);
@@ -49,5 +48,5 @@ module.exports.destroy = async function(req, res){
             message: "Internal Server Error"
         });
     }
-    
+     
 }
